+++ BossMod/Modules/Endwalker/Savage/P1SErichthonios/Flails.cs
@@ -47,8 +47,8 @@ class Flails : BossComponent
         if (!_detectSecond)
             return;
 
-        var weaponsBall = Module.Enemies(OID.FlailI);
-        var weaponsChakram = Module.Enemies(OID.FlailO);
+        var weaponsBall = Module.Enemies((uint)OID.FlailI);
+        var weaponsChakram = Module.Enemies((uint)OID.FlailO);
         if (weaponsBall.Count + weaponsChakram.Count > 0)
         {
             _detectSecond = false;
