+++ BossMod/Modules/Endwalker/Savage/P2SHippokampos/Dissociation.cs
@@ -7,7 +7,7 @@ class Dissociation(BossModule module) : BossComponent(module)
 
     public override void AddHints(int slot, Actor actor, TextHints hints)
     {
-        var head = Module.Enemies(OID.DissociatedHead).FirstOrDefault();
+        var head = Module.Enemies((uint)OID.DissociatedHead).FirstOrDefault();
         if (_shape == null || head == null || Module.InBounds(head.Position))
             return; // inactive or head not teleported yet
 
@@ -19,7 +19,7 @@ class Dissociation(BossModule module) : BossComponent(module)
 
     public override void DrawArenaBackground(int pcSlot, Actor pc)
     {
-        var head = Module.Enemies(OID.DissociatedHead).FirstOrDefault();
+        var head = Module.Enemies((uint)OID.DissociatedHead).FirstOrDefault();
         if (_shape == null || head == null || Module.InBounds(head.Position))
             return; // inactive or head not teleported yet
