+++ BossMod/Modules/RealmReborn/Extreme/Ex1Ultima/Ex1UltimaAI.cs
@@ -9,7 +9,7 @@ class Ex1UltimaAI(BossModule module) : BossComponent(module)
 
     public override void AddAIHints(int slot, Actor actor, PartyRolesConfig.Assignment assignment, AIHints hints)
     {
-        if (Module.PrimaryActor.TargetID != actor.InstanceID && hints.ForbiddenZones.Count == 1 && !Module.Enemies(OID.MagitekBit).Any(a => !a.IsDead)) // for non-mt, there is always a cleave
+        if (Module.PrimaryActor.TargetID != actor.InstanceID && hints.ForbiddenZones.Count == 1 && !Module.Enemies((uint)OID.MagitekBit).Any(a => !a.IsDead)) // for non-mt, there is always a cleave
         {
             // default positions: tank boss at the edge facing N, OT south of boss, M1/M2 to the left/right (so that they can slightly adjust for positionals), H1/H2/R1/R2 to S outside ceruleum vent range, all spread somewhat to avoid homing lasers
             // when tanks need to swap, OT moves between boss and MT and taunts; OT needs to ignore diffractive lasers at this point
